Updated pointlogger.js application to support new timestamp options

+++ apps/Point-monitor/pointlogger.js
@@ -3,11 +3,15 @@ const mqtt = require('mqtt');
 const fs = require('fs'); 
 const child_process = require('child_process');
 
-const appVersion = '1.03.003';
+const appVersion = '1.03.004';
 const devStates = new Map();
 const nameMap = new Map();
 const msToSet = new Map();
 const monitorSetRequestDelay = 2000;
+const zeroPad = (num, places) => String(num).padStart(places, '0');
+let timeZoneOffset  = new Date().getTimezoneOffset();
+const tzString = `${(timeZoneOffset > 0) ? '-' : '+'}${zeroPad (timeZoneOffset/60,2)}${(timeZoneOffset % 60 !=0) ? '30' : '00'}`;
+timeZoneOffset = timeZoneOffset * 60000; // Must be in msec for the Date.setTime method
 
 // Reords are timestamped in a combined ISO 8601 format in GMT time
 // YYYY-MM-DDTHH:MM:SS.mmmZ.  ImportReadyTimestamps use the javascript
@@ -28,25 +32,29 @@ let filePrefix = 'A';
 
 // Edit monitorSpecs object array according to your requirements to monitor the target data points of interest
 // Review the IAP/MQ documentation on the use of event: true.  This can only be used in DMM managed systems.
-// the ms object below is used to set the monitoring objects of multiple points with one IAP/MQ request
+// the ms object below is used to set the monitoring objects of multiple points with one IAP/MQ request (9000011503020467 6K min simulator)
+// TODO: monitorSpecs must be updated to reference points intended for logging.  Polling rates must be scaled to
+// the scope of the system.
 var monitorSpecs = [
-    {
-        pid: '9000011503020465',
-        ptPath: 'device/0',
-        ms:{ 
-            nvoPF: { monitor: {rate: 60, report: 'any', inFeedback:false, event:false}},
-            nvoI: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}}, 
-            nvoU_F: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}},
-            nvoP: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}},
-            nvoEPpos: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}}
-            //nvoPF_2: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}},
-            //nvoI_2: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}},
-            //nvoU_F_2: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}},
-            //nvoP_2: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}},
-            //nvoEPapos: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}}
-        }
-    }
-    
+    {	
+        pid: '9000011503020467',
+            ptPath: 'device/0',
+            ms:{ 
+                nvoTotalPower: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}},
+                nvoEPpos: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}}
+            }
+    },
+    {	
+        pid: '9FFFFF0501840450',
+            ptPath: 'LightSensor/0',
+            ms:{ 
+                nvoLuxLevel: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}},
+            },
+            ptPath: 'TempSensor/0',
+            ms:{ 
+                nvoHVACTemp: {monitor: {rate: 60, report: 'any', inFeedback:false, event:false}},
+            }
+    } 
 ];
 // environ returns the value of a named environment variable, if it exists, or returns the default value otherwise.
 function environ(variable, defaultValue) {
@@ -58,6 +66,9 @@ const client = mqtt.connect(`mqtt://${environ('DEV_TARGET', '127.0.0.1')}:1883`)
 
 let logTruncate = true;
 let rateOverride = -1;  // -2 means DLA managed polling. -1 means use monitor spec value. >= 0 set as the monitor rate
+                        // -3 means log to single file without staging to transfer folder.  -4 mean monitor data events without logging.
+                        // use -4 to establish the eps base line.
+let useLocalTz = 0;     // Added in 1.03.004 to store times in local time.  Third cmdline argument 1|0                
 let  args = process.argv.slice(2);  
 if (args.length === 1)
     rateOverride = parseInt(args[0]);
@@ -66,7 +77,11 @@ if (args.length === 2){
     rateOverride = parseInt(args[0]);
     delayStart = JSON.parse(args[1]) ;
 }
-
+if (args.length === 3){
+    rateOverride = parseInt(args[0]);
+    delayStart = JSON.parse(args[1]) ;
+    useLocalTz = parseInt(args[2]);  // Set to 1 if you want log timestamps to reflect local time and NOT UTC
+}
 // The first argument is used to override the monitorSpec polling rate. 
 if (rateOverride >= 0) {
     monitorSpecs.forEach(target => {
@@ -88,10 +103,10 @@ if (delayStart) {
 // Determine available file system assets.  Using SD card on SmartServer if available
 // See Integration TODO:
 const eloggerDataDir = process.env.hasOwnProperty('APOLLO_DATA') ? '/media/sdcard' : '.';
-if (rateOverride === -2)
+if (rateOverride === -2 || rateOverride <= -3)
     console.log ('An active DLA file must be used to start monitoring');
 
-// Integration TODO:
+// Integration steps on your target TODO:
 // 1. On Apollo target: sudo mkdir /media/sdcard/eloggerdata
 // 2. sudo mkdir /media/sdcard/transfer
 // 3. sudo chown apollo:apollo /media/sdcard/eloggerdata
@@ -116,18 +131,46 @@ function reportDeviceState () {
     }
     console.log (`Up Devices - ${upDeviceList}`);
     console.log (`Down Devices - ${downDeviceList}`);
-}
-// Scheduling copy to the transfer folder every 5m (300s)
+} 
+let lastCount = 0;
+let logRate = 0.0;
+
 setInterval(() => {
     let nowTs = new Date();
+    let recDelta;
+    let logRateNow = 0.0;
+    recDelta = stagedRecords - lastCount;
+    lastCount = stagedRecords;
+    logRateNow = recDelta / 10.0;
+    if (Math.abs (logRateNow - logRate) > 2) {
+        console.log(`${nowTs.toLocaleString()} Log Rate: ${logRateNow} per sec`); 
+        logRate = logRateNow;
+    } 
+
+}, 10000);
+// Scheduling copy to the transfer folder every 5m (300s)
+
+setInterval(() => {
+    let nowTs = new Date();
+    let localTs = new Date();
+
+    // No data, no transfer staging required
+    if (stagedRecords == 0 || rateOverride <= -3)
+        return; 
+    // 1.03.004 - Change to support a human readable TS ${sid}-yyyymmdd_hh_mm_ss
+    // using local time
+    let fileTs = `${localTs.getFullYear()}${zeroPad(localTs.getMonth() + 1,2)}${zeroPad(localTs.getDate(),2)}_${zeroPad(localTs.getHours(),2)}_` +
+            `${zeroPad(localTs.getMinutes(),2)}_${zeroPad(localTs.getSeconds(),2)}`;
+
     fs.copyFile(`${eloggerDataDir}/eloggerdata/${filePrefix}point-data.csv`, 
-        `${eloggerDataDir}/transfer/${sid}-${parseInt(nowTs/1000)}.csv`,
+        `${eloggerDataDir}/transfer/${sid}-${fileTs}.csv`,
         (err) => {
             if (err) 
                 console.error(err);
             else {
                 console.log(`${nowTs.toLocaleString()} Transfer Staged. Monitored Devices: ${devStates.size}. Staged Records: ${stagedRecords}`);
                 stagedRecords = 0;
+                lastCount = 0;
                 if (filePrefix === 'A')
                     fs.truncate(`${eloggerDataDir}/eloggerdata/Bpoint-data.csv`, 
                         (err) => {
@@ -232,7 +275,7 @@ function handleDeviceSts (devHandle, stsMsg, monSpecIndex) {
             console.log(`${nowTs.toLocaleString()} - Device: ${devHandle} (S/N: ${stsMsg.addr.domain[0].subnet}/${stsMsg.addr.domain[0].nodeId}) is ${stsMsg.health}.`);
             return false;
         }
-        if (rateOverride == -2 || !setMs) {
+        if (rateOverride == -2 || rateOverride <= -3 || !setMs) {
             if (!setMs)
                 console.log(`${nowTs.toLocaleString()} - Monitor Object already set.`);
             return false;
@@ -327,48 +370,64 @@ client.on('message', (topic, message) => {
             devHandle =  payload.topic.split('/')[6];
             // Only looking for data events from targed devices. Just sending to the console in this example
             if (devStates.has(devHandle)) {
+
+                if (rateOverride == -4)
+                    return;
                 dpState = devStates.get(devHandle) === 'normal' ? 'ONLINE' : 'OFFLINE';
                 pointPath = `${nameMap.get(devHandle)}/${payload.message.split('/')[0]}`
                 // Use the Timestamp from the inside data event  MQTT client process the message
                 // This timestamp is in UTC. The back end should convert to the site's local timezone
                 // Added a constant to force conversion to the local time in release 1.30.003 
-                // importReadyTimeSt
+                // importReadyTimeSt/.  1.30.004 added support for a third cmdline argument that when set to 1
+                // will log records in the local Time using ISO time notation: YYYY-MM-DDTHH:MM:SS:ms[+|-]hh:mm (offset to the local TZ) 
 
                 if (importReadyTimestamps) {
                     let timeStamp = `${payload.ts}`;
                     let localTs = new Date(timeStamp);
-                    logRecord = `${localTs.valueOf()},\"${pointPath}\",\"${dpState}"\,\"`; 
+                    logRecord = `${localTs.valueOf()},\"${devHandle}\",\"${pointPath}\",\"${dpState}"\,\"`; 
                 } else { 
-                    logRecord = `${payload.ts},\"${pointPath}\",\"${dpState}"\,\"`; 
+                    if (!useLocalTz)
+                        logRecord = `${payload.ts},\"${devHandle}\",\"${pointPath}\",\"${dpState}"\,\"`; 
+                    else {
+                        let timeStamp = `${payload.ts}`;
+                        const localTs = new Date(timeStamp);
+                        // Adjust for TZ offset
+                        localTs.setTime(localTs.valueOf() - timeZoneOffset);
+                        // Using precomputed tzString offset 
+                        logRecord = `${localTs.toISOString().substr(0,23)}${tzString},\"${devHandle}\",\"${pointPath}\",\"${dpState}"\,\"`; 
+                    }
                 }    
 
                 // The application has knowledge of the payload data to only log the relavent values.  The SNVT_trans_table values are using
-                // 3 or 4 floats depending on the network variable.  If a data event 
-                if (payload.message === 'nvoP/value') 
-                    logRecord += `${payload.data.point[0]} ${payload.data.point[1]} ${payload.data.point[2]} ${payload.data.point[3]}\"`;
-                else if (payload.message === 'nvoP_2/value') 
-                    logRecord += `${payload.data.point[0]} ${payload.data.point[1]} ${payload.data.point[2]} ${payload.data.point[3]}\"`;
-                else if  (payload.message === 'nvoI/value')
-                    logRecord += `${payload.data.point[0]} ${payload.data.point[1]} ${payload.data.point[2]}\"`;       
-                else if  (payload.message === 'nvoI_2/value')
-                    logRecord += `${payload.data.point[0]} ${payload.data.point[1]} ${payload.data.point[2]}\"`;        
-                else if  (payload.message === 'nvoU_F/value') 
-                    logRecord += `${payload.data.point[0]} ${payload.data.point[1]} ${payload.data.point[2]}\"`;           
+                // 3 or 4 floats depending on the network variable.  TODO: Modify the payload.message checks to support the datapoints you have defined
+                // in the monitorSpec.
+
+                if (payload.message === 'nvoTotalPower/value') 
+                    logRecord += `${payload.data}\"`;
+                else if (payload.message === 'nvoPower2/value') 
+                    logRecord += `${payload.data.point[0]}\"`;
+                else if  (payload.message === 'nvoEPsumZZ/value')
+                    logRecord += `${payload.data}\"`; 
+                else if  (payload.message === 'nvoHVACTemp/value')
+                    logRecord += `${payload.data}\"`;
+                else if  (payload.message === 'nvoLuxLevel/value') 
+                    logRecord += `${payload.data}\"`;
                 else if  (payload.message === 'nvoU_F_2/value') 
-                    logRecord += `${payload.data.point[0]} ${payload.data.point[1]} ${payload.data.point[2]}\"`;          
+                    return; //logRecord += `${payload.data.point[0]} ${payload.data.point[1]} ${payload.data.point[2]}\"`;          
                 else if  (payload.message === 'nvoPF/value') 
                     logRecord += `${payload.data.point[0]} ${payload.data.point[1]} ${payload.data.point[2]} ${payload.data.point[3]}\"`;
                 else if  (payload.message === 'nvoPF_2/value') 
-                    logRecord += `${payload.data.point[0]} ${payload.data.point[1]} ${payload.data.point[2]} ${payload.data.point[3]}\"`;
+                    return; ///logRecord += `${payload.data.point[0]} ${payload.data.point[1]} ${payload.data.point[2]} ${payload.data.point[3]}\"`;
                 else if  (payload.message === 'nvoEPpos/value')
                     logRecord += `${payload.data}\"`; 
                 else if  (payload.message === 'nvoEPapos/value')
-                    logRecord += `${payload.data}\"`; 
+                    return; //logRecord += `${payload.data}\"`; 
                 else if (payload.message === 'nvoCounterData/value')    
-                    logRecord += `${payload.data.counter[0]} ${payload.data.counter[1]} ${payload.data.counter[2]} ${payload.data.counter[3]} ${payload.data.faultCounter}\"`;
+                    return; //logRecord += `${payload.data.counter[0]} ${payload.data.counter[1]} ${payload.data.counter[2]} ${payload.data.counter[3]} ${payload.data.faultCounter}\"`;
                 else // The application could be designed to ONLY filter on the 5 key network variables, and might abort loging the data at this point
-                    logRecord += `${JSON.stringify(payload.data)}\"`;
-                
+                    return; //logRecord += `${JSON.stringify(payload.data)}\"`;
+                ++recordCount;
+                ++stagedRecords;
                 fs.writeFile(`${eloggerDataDir}/eloggerdata/${filePrefix}point-data.csv`, `${logRecord}\r\n`, logTruncate ? {flag: 'w'} :
                     {flag:'a+'}, (err) => {
                         if (err) {
@@ -377,8 +436,7 @@ client.on('message', (topic, message) => {
                         }
                 }); 
                 logTruncate = false; 
-                ++recordCount;
-                ++stagedRecords;
+
                 if (testLimit !== -1 && recordCount >= testLimit) {
                     client.end(true);
                     console.log (`Test completed. ${recordCount} data events captured`);
